requirements.txt
Flask
Flask-Cors
requests
PyJWT
werkzeug